requirements.txt
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.29.0
pyzmq>=25.0.0
rich>=13.0.0
numpy>=1.26.0
pyyaml>=6.0.0
